re-org the component files related to Calendar component
