+++ src/component/Calendar/index.js
@@ -1,0 +1,3 @@
+import Calendar from './Calendar';
+
+export default Calendar;
